# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Meyora API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/integrations/gmail": {
   "get": {
    "tags": [
     "integrations"
    ],
    "summary": "Gmail Status",
    "operationId": "gmail_status_api_integrations_gmail_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GmailStatus"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/integrations/gmail/connect": {
   "post": {
    "tags": [
     "integrations"
    ],
    "summary": "Gmail Connect",
    "operationId": "gmail_connect_api_integrations_gmail_connect_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/integrations/gmail/callback": {
   "get": {
    "tags": [
     "integrations"
    ],
    "summary": "Gmail Callback",
    "operationId": "gmail_callback_api_integrations_gmail_callback_get",
    "parameters": [
     {
      "name": "code",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Code"
      }
     },
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "State"
      }
     },
     {
      "name": "error",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Error"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/integrations/gmail/{account_id}/sync": {
   "post": {
    "tags": [
     "integrations"
    ],
    "summary": "Gmail Sync Now",
    "operationId": "gmail_sync_now_api_integrations_gmail__account_id__sync_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "account_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Account Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MailAccountOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/integrations/gmail/{account_id}": {
   "delete": {
    "tags": [
     "integrations"
    ],
    "summary": "Gmail Disconnect",
    "operationId": "gmail_disconnect_api_integrations_gmail__account_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "account_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Account Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/emails/threads": {
   "get": {
    "tags": [
     "emails"
    ],
    "summary": "List Threads",
    "operationId": "list_threads_api_emails_threads_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "contact_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Contact Id"
      }
     },
     {
      "name": "company_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company Id"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 10000,
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 25,
       "title": "Page Size"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 200
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 50,
         "pattern": "^-?[a-z_]+$"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sort"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_ThreadOut_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/emails/threads/{thread_id}": {
   "get": {
    "tags": [
     "emails"
    ],
    "summary": "Get Thread",
    "operationId": "get_thread_api_emails_threads__thread_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "thread_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Thread Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ThreadDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "GmailStatus": {
    "properties": {
     "enabled": {
      "type": "boolean",
      "title": "Enabled"
     },
     "accounts": {
      "items": {
       "$ref": "#/components/schemas/MailAccountOut"
      },
      "type": "array",
      "title": "Accounts"
     }
    },
    "type": "object",
    "required": [
     "enabled",
     "accounts"
    ],
    "title": "GmailStatus"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MailAccountOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "user_id": {
      "type": "string",
      "format": "uuid",
      "title": "User Id"
     },
     "email_address": {
      "type": "string",
      "title": "Email Address"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "last_sync_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Sync At"
     },
     "next_sync_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Sync At"
     },
     "last_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Error"
     },
     "messages_synced": {
      "type": "integer",
      "title": "Messages Synced"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "user_id",
     "email_address",
     "status",
     "last_sync_at",
     "next_sync_at",
     "last_error",
     "messages_synced",
     "created_at"
    ],
    "title": "MailAccountOut"
   },
...
   "Page_ThreadOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ThreadOut"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[ThreadOut]"
   },
...
   "ThreadDetail": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "subject": {
      "type": "string",
      "title": "Subject"
     },
     "snippet": {
      "type": "string",
      "title": "Snippet"
     },
     "participants": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Participants"
     },
     "contact_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Contact Ids"
     },
     "company_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Company Ids"
     },
     "message_count": {
      "type": "integer",
      "title": "Message Count"
     },
     "last_message_at": {
      "type": "string",
      "format": "date-time",
      "title": "Last Message At"
     },
     "messages": {
      "items": {
       "$ref": "#/components/schemas/app__api__emails__MessageOut"
      },
      "type": "array",
      "title": "Messages"
     }
    },
    "type": "object",
    "required": [
     "id",
     "subject",
     "snippet",
     "participants",
     "contact_ids",
     "company_ids",
     "message_count",
     "last_message_at",
     "messages"
    ],
    "title": "ThreadDetail"
   },
   "ThreadOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "subject": {
      "type": "string",
      "title": "Subject"
     },
     "snippet": {
      "type": "string",
      "title": "Snippet"
     },
     "participants": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Participants"
     },
     "contact_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Contact Ids"
     },
     "company_ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Company Ids"
     },
     "message_count": {
      "type": "integer",
      "title": "Message Count"
     },
     "last_message_at": {
      "type": "string",
      "format": "date-time",
      "title": "Last Message At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "subject",
     "snippet",
     "participants",
     "contact_ids",
     "company_ids",
     "message_count",
     "last_message_at"
    ],
    "title": "ThreadOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "app__api__emails__MessageOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "direction": {
      "type": "string",
      "title": "Direction"
     },
     "from_email": {
      "type": "string",
      "title": "From Email"
     },
     "from_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "From Name"
     },
     "to": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "To"
     },
     "cc": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Cc"
     },
     "subject": {
      "type": "string",
      "title": "Subject"
     },
     "body_text": {
      "type": "string",
      "title": "Body Text"
     },
     "has_attachments": {
      "type": "boolean",
      "title": "Has Attachments"
     },
     "sent_at": {
      "type": "string",
      "format": "date-time",
      "title": "Sent At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "direction",
     "from_email",
     "from_name",
     "to",
     "cc",
     "subject",
     "body_text",
     "has_attachments",
     "sent_at"
    ],
    "title": "MessageOut"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}